BusinessLogic/requirements.txt
lxml
pandas
requests
beautifulsoup4
pymongo
python-decouple
scipy